Damgard-Fujisaki tests split into two files.
